refactor: added convertField object for converting messages
