+++ src/util/convertToDiscord/convertMessageToDiscord.ts
@@ -1,19 +1,6 @@
-import {
-  APIEmbed,
-  APIUser,
-  MessageType,
-  APIMessage,
-  InteractionType,
-  APIMessageInteraction
-} from 'discord-api-types/v10';
+import { MessageType, APIMessage, ChannelType } from 'discord-api-types/v10';
 import { Message } from '@graphql/graphql';
-import { getAvatarId } from '@util/convertToDiscord/getAvatarId';
-
-const getIdFromUrl = (avatarUrl: string) => {
-  const id = avatarUrl.split('/')[4] ?? null;
-
-  return id;
-};
+import { convertField } from '@util/convertToDiscord/convertField';
 
 export const convertMessageToDiscord = (message: Message): APIMessage & { isGuest: boolean } => ({
   id: message.id,
@@ -28,127 +15,39 @@ export const convertMessageToDiscord = (message: Message): APIMessage & { isGues
   flags: message.flags ?? 0,
   isGuest: message.isGuest,
 
-  author: {
-    id:
-      message.isGuest && message.author.bot
-        ? getIdFromUrl(message.author.avatarUrl)
-        : message.author.id,
-    bot: message.author.bot,
-    username: message.author.name,
-    avatar: getAvatarId(message.author.avatarUrl),
-    global_name: message.author.name,
-    system: message.author.system,
-    flags: message.author.flags ?? 0,
-    discriminator: message.author.discrim
-  },
+  author: convertField.author(message),
 
-  attachments: message.attachments.map(a => ({
-    id: '0',
-    url: a.url,
-    height: a.height,
-    filename: a.filename,
-    size: a.size,
-    proxy_url: a.url,
-    width: a.width
-  })),
+  attachments: convertField.attachments(message.attachments),
 
-  embeds: message.embeds
-    ? (message.embeds.map(e => ({
-        author: {
-          name: e.author?.name,
-          icon_url: e.author?.icon,
-          proxy_icon_url: e.author?.proxyIconUrl,
-          url: e.author?.url
-        },
+  embeds: convertField.embeds(message.embeds),
 
-        color: e.color,
-        description: e.description,
+  reactions: convertField.reactions(message.reactions),
 
-        footer: {
-          text: e.footer?.text,
-          icon_url: e.footer?.proxyIconUrl,
-          proxy_icon_url: e.footer?.proxyIconUrl
-        },
+  mentions: convertField.mentions(message.mentions),
 
-        image:
-          {
-            url: e.image?.url,
-            height: e.image?.height,
-            proxy_url: e.image?.proxyUrl,
-            width: e.image?.width
-          } ?? undefined,
+  interaction: convertField.interaction(message.interaction),
 
-        provider: e.provider,
+  referenced_message: message.referencedMessage
+    ? convertMessageToDiscord(message.referencedMessage)
+    : undefined,
 
-        thumbnail:
-          {
-            url: e.thumbnail?.url,
-            height: e.thumbnail?.height,
-            proxy_url: e.thumbnail?.proxyUrl,
-            width: e.thumbnail?.width
-          } ?? undefined,
+  message_reference: message.messageReference
+    ? {
+        channel_id: message.messageReference.channelId,
+        guild_id: message.messageReference.guildId as string,
+        message_id: message.messageReference.messageId as string
+      }
+    : undefined,
 
-        timestamp: e.timestamp,
-        title: e.title,
-        url: e.url,
-
-        fields: e.fields,
-
-        video:
-          {
-            height: e.video?.height,
-            proxy_url: e.video?.proxyUrl,
-            url: e.video?.url,
-            width: e.video?.width
-          } ?? undefined
-      })) as APIEmbed[])
-    : [],
-
-  reactions: message.reactions
-    ? message.reactions.map(r => ({
-        count: r.count as number,
-
-        emoji: {
-          id: r.emojiId as string,
-          name: r.emojiName as string,
-          animated: r.animated as boolean
-        },
-
-        me: r.me
-      }))
-    : [],
-
-  mentions: message.mentions
-    ? (message.mentions.map(m => ({
-        id: m.id,
-        guild_id: m.id,
-        type: m.type,
-        name: m.name,
-        global_name: m.name,
-        discriminator: '0000',
-        username: m.name,
-        avatar: ''
-      })) as APIUser[])
-    : [],
-
-  interaction: message.interaction
-    ? ({
-        id: message.interaction.id,
-        name: message.interaction.name,
-
-        user: {
-          avatar: getAvatarId(message.interaction.user.avatarUrl),
-          discriminator: '0000',
-          global_name: message.interaction.user.username,
-          username: message.interaction.user.username,
-          id: message.interaction.user.bot
-            ? getIdFromUrl(message.interaction.user.avatarUrl)
-            : message.interaction.user.id,
-          bot: message.interaction.user.bot
-        },
-
-        type: InteractionType.MessageComponent
-      } as APIMessageInteraction)
+  thread: message.thread
+    ? {
+        id: message.thread.id,
+        name: message.thread.name,
+        message_count: message.thread.messageCount,
+        position: 0,
+        type: ChannelType.PublicThread,
+        applied_tags: []
+      }
     : undefined,
 
   mention_roles: [],
